% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmnr-formula.R
\name{bmnr-formula}
\alias{bmnr-formula}
\alias{parse_bmnr_formula}
\title{Parse \code{\link{bmnr}} Formula}
\usage{
parse_bmnr_formula(object)
}
\arguments{
\item{object}{formula. User supplied model description.}
}
\description{
Description of how to specify models with coordinate metadta.
}
\section{Extended Formula}{

For more information on extending \code{formula}, see the \code{Formula} package but
be warned about \url{https://github.com/rstudio/rstudio/issues/12409}.

Specific a standard LHS as one would with \code{\link[stats:lm]{stats::lm()}} using \code{cbind}.
\itemize{
\item \code{cbind(y01, y02) ~ x01 + x02}
\item \code{cbind(y01, y02) ~ .}
}

The RHS is split into predictors and coordinate information split into two
parts using \code{"|"}

For example,
\itemize{
\item \code{cbind(y01, y02) ~ x01 + x02 + x03 | coord_x + coord_y}
}
}

