% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmnr-package.R
\docType{package}
\name{bmnr-package}
\alias{bmnr-package}
\title{Bayesian Matrix Normal Regression Package}
\arguments{
\item{object}{Object that inherits from \code{\link{bmnrfit}}.}

\item{new_data}{data frame. Data used to create a design matrices.
Setting \code{new_data = NULL} utilises the pre-trained data (special case).}
}
\description{
Fits a Matrix normal distribution where the mean is a linear predictor,
covariance across rows is assumed Gaussian process (GP) and
covariance across columns is estimated.
}
\section{Model description}{

Suppose we observe two matrices,
\itemize{
\item \eqn{Y}: \eqn{n_s} observations for \eqn{n_y} dependent variables;
\item \eqn{X}: \eqn{n_s} observations for \eqn{n_x} independent variables.
}

Our model assumes
\deqn{
    Y \sim MN_{n_s, n_y}(XB, \Sigma_s, \Sigma_y),
  }
where
\itemize{
\item \eqn{B}, \eqn{n_x + 1} by \eqn{n_y} regression coefficient matrix;
\item \eqn{\Sigma_s}, (across row) covariance matrix, determined by a Gaussian
process (function of \eqn{n_s} coordinates and hyperparameters);
\item \eqn{\Sigma_y}, (across column) covariance matrix.
}
}

\section{Prior Information}{

It is assumed, \emph{a priori}, that
\itemize{
\item \eqn{B_{ij} \sim N(0, \tau_B)} where \eqn{\tau_B =} \code{regr_prec}.
\item \strong{TODO}. Complete prior section.
}
}

\references{
Stan Development Team (2023). RStan: the R interface to Stan.
R package version 2.32.3. \url{https://mc-stan.org}
}
