% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmnr-autoplot.R
\name{bmnr-plot}
\alias{bmnr-plot}
\alias{autoplot.bmnrfit}
\title{Plot Functionality for \code{bmnr}}
\usage{
\method{autoplot}{bmnrfit}(object, pars, type, ...)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmnrfit}}.}

\item{pars}{character. Parameter names, with optional subsetting.}

\item{type}{choice. Either \code{density}, \code{trace} or \code{acf}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Methods for the \code{\link[ggplot2:autoplot]{ggplot2::autoplot}} generic.
}
