% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmnr-predict.R
\name{bmnr-predict}
\alias{bmnr-predict}
\alias{predict.bmnrfit}
\alias{predict.bmnrfit_mvrnorm}
\alias{extract_posterior_predictive_pars_bmnrfit}
\alias{convert_raw_predictions_to_output}
\title{Prediction of \code{bmnr} Models}
\usage{
\method{predict}{bmnrfit}(object, new_data, ..., type = NULL)

\method{predict}{bmnrfit_mvrnorm}(object, new_data, ..., type = NULL)

extract_posterior_predictive_pars_bmnrfit(object, new_data)

convert_raw_predictions_to_output(x, type, .names = NULL)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmnrfit}}.}

\item{new_data}{data frame. Data used to create a design matrices.
Setting \code{new_data = NULL} utilises the pre-trained data (special case).}

\item{...}{Extra arguments passed to \code{\link[mvtnorm:Mvnorm]{mvtnorm::rmvnorm()}}.}

\item{type}{choice. Return type, see \strong{Value} section for more information}

\item{x}{array. Raw MCMC output as described in \strong{Value} section.}

\item{.names}{character. Names each dependent variable to be predicted.}
}
\value{
The helper \code{extract_posterior_predictive_pars()} returns a list of the
distribution parameters in the form of a \verb{$flattened_mean} and \verb{$covar}.
Output of the \code{predict} method is dictated by the \code{type} argument:
\itemize{
\item \code{summary}

Tibble representing posterior predictive summaries;
one row per supplied observation from \code{new_data}.
\item \code{numeric}

Tibble with raw MCMC objects stored in a single named list column;
one row per supplied observation from \code{new_data}.
\item \code{raw}

MCMC array of dimension \code{c(n_iter, n_chains, nrow(new_data), n_y)}.
}
}
\description{
Prediction of \code{bmnr} Models
}
\section{Posterior Predictive Distribution}{

TODO. Convert from handwritten notes to Markdown
}

\keyword{internal}
