% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{bmnr-render-internal}
\alias{bmnr-render-internal}
\alias{.set_bmnr_render_options}
\title{Functions used in Analysis Report}
\usage{
.set_bmnr_render_options(object, is_child, fig.asp)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmnrfit}}.}

\item{is_child}{logical. Some options are only set when this is \code{FALSE}.}

\item{fig.asp}{numeric. Default aspect ratio per \strong{panel}, not per plot.}
}
\description{
Functions not to be used anywhere other than \code{bmnr-analysis.qmd}.
}
