% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmnr.R
\name{bmnr}
\alias{bmnr}
\alias{bmnr.bmnr_sim}
\alias{bmnr.formula}
\title{Bayesian Matrix Normal Regression}
\usage{
bmnr(object, ..., prior)

\method{bmnr}{bmnr_sim}(object, ..., prior)

\method{bmnr}{formula}(object, coord_formula, data, ..., prior)
}
\arguments{
\item{object}{Typically a formula (with \code{data}) but can be \code{\link{bmnr_sim}}.}

\item{...}{Extra arguments passed to \link[rstan:stanmodel-method-sampling]{rstan::sampling}.}

\item{prior}{list. Required prior information; see \code{\link[=bmnr_prior]{bmnr_prior()}}.}

\item{coord_formula}{RHS only formula. Specifies coordinate data.\
\strong{Note}. Future versions will pull this argument into \code{formula}.}

\item{data}{data.frame. Training data that contains all \code{formula} terms.}
}
\description{
Bayesian Matrix Normal Regression
}
