% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmnr-yx.R
\name{bmnr_yx}
\alias{bmnr_yx}
\title{Implementation of BMNR Model}
\usage{
bmnr_yx(y, x, coords, ..., prior)
}
\arguments{
\item{y}{matrix Dependent variables.}

\item{x}{matrix. Explanatory variables.}

\item{coords}{matrix. Gaussian process inputs.}

\item{...}{Extra arguments passed to \link[rstan:stanmodel-method-sampling]{rstan::sampling}.}

\item{prior}{list. Required prior information; see \code{\link[=bmnr_prior]{bmnr_prior()}}.}
}
\description{
Implementation of BMNR Model
}
\keyword{internal}
