% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmnr-extract-data-frame.R
\name{extract_simstudy_true_values}
\alias{extract_simstudy_true_values}
\title{Get True Value from a Simulation Study}
\usage{
extract_simstudy_true_values(object, pars)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmnrfit_simstudy}}.}

\item{pars}{character. Parameter names, with optional subsetting.}
}
\description{
Get True Value from a Simulation Study
}
\keyword{internal}
