% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmnr-extract.R
\name{extract_to_data_frame}
\alias{extract_to_data_frame}
\alias{extract_to_data_frame,stanfit-method}
\alias{extract_to_data_frame,bmnrfit_simstudy-method}
\title{Extract Samples to Data Frame}
\usage{
extract_to_data_frame(object, pars, ...)

\S4method{extract_to_data_frame}{stanfit}(object, pars, ...)

\S4method{extract_to_data_frame}{bmnrfit_simstudy}(object, pars, ...)
}
\arguments{
\item{object}{Object that inherits from \code{\link[rstan:stanfit-class]{stanfit}}.}

\item{pars}{character. Parameter names, with optional subsetting.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Wrapper for \code{\link[rstan:stanfit-method-extract]{rstan::extract()}}; converts array of samples to tidy format.
}
\note{
Applying this method to an object of class \code{\link{bmnrfit_simstudy}} adds
an extra \code{.truth} column as appropriate.
}
