% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmnr-names.R
\name{get_names_from_bmnrfit}
\alias{get_names_from_bmnrfit}
\title{Get Names from \code{\link{bmnr}}}
\usage{
get_names_from_bmnrfit(object)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmnrfit}}.}
}
\value{
List with character vector elements representing original names.
}
\description{
Get Names from \code{\link{bmnr}}
}
