% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{get_stan_controls}
\alias{get_stan_controls}
\title{Extract Algorithm Controls from \code{stanfit}}
\usage{
get_stan_controls(object)
}
\arguments{
\item{object}{Object that inherits from \code{stanfit}.}
}
\description{
Extract Algorithm Controls from \code{stanfit}
}
