% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmnr-extract.R
\name{mcmcarray_to_data_frame}
\alias{mcmcarray_to_data_frame}
\title{Convert Stan's MCMC output to Data Frame}
\usage{
mcmcarray_to_data_frame(mcmcarray)
}
\arguments{
\item{mcmcarray}{array. Dimensions should be iterations, chains and label.}
}
\description{
Convert Stan's MCMC output to Data Frame
}
\keyword{internal}
