% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_unif_spatiotemporal_coords}
\alias{sim_unif_spatiotemporal_coords}
\title{Simulate Spatiotemporal Coordinates}
\usage{
sim_unif_spatiotemporal_coords(n_locations, n_years)
}
\arguments{
\item{n_locations}{integer. Number of locations from \eqn{U(0, 1)^2}.}

\item{n_years}{integer. Number of years to simulate}
}
\value{
matrix.
Spatial coordinates are \eqn{[0,1]} and temporal units are years.
}
\description{
Get sampling coordinates for \code{n_locations} with a consistent quarterly
schedule over \code{n_years} years.
}
\keyword{internal}
