% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_bmnr}
\alias{simulate_bmnr}
\alias{bmnr-sim}
\alias{example_simulate_bmnr}
\title{Simulate \strong{Bayesian Matrix Normal Regression}}
\usage{
simulate_bmnr(n_y, n_x, regr, covar_y, coords, gp_func, ...)

example_simulate_bmnr(...)
}
\arguments{
\item{n_y}{integer. Number of dependent variables.}

\item{n_x}{integer. Number of independent variables.}

\item{regr}{matrix (\eqn{n_x + 1} by \eqn{n_y}). Regression coefficients.}

\item{covar_y}{\eqn{\Sigma_y} matrix. Covariance across columns.}

\item{coords, gp_func, ...}{Determines the covariance across rows, specifically
\eqn{\Sigma_s = } \code{gp_func(coords, ...)}.}
}
\value{
List with class (\code{bmnr_sim}) containing \verb{$data} and \verb{$params}.
}
\description{
Simulates from a BMNR model.
}
\section{Functions}{
\itemize{
\item \code{example_simulate_bmnr()}: Default arguments for quick simulation.

}}
\note{
Setting \code{gp_func = NULL} implies independence across rows.
}
\section{Model description}{

Suppose we observe two matrices,
\itemize{
\item \eqn{Y}: \eqn{n_s} observations for \eqn{n_y} dependent variables;
\item \eqn{X}: \eqn{n_s} observations for \eqn{n_x} independent variables.
}

Our model assumes
\deqn{
    Y \sim MN_{n_s, n_y}(XB, \Sigma_s, \Sigma_y),
  }
where
\itemize{
\item \eqn{B}, \eqn{n_x + 1} by \eqn{n_y} regression coefficient matrix;
\item \eqn{\Sigma_s}, (across row) covariance matrix, determined by a Gaussian
process (function of \eqn{n_s} coordinates and hyperparameters);
\item \eqn{\Sigma_y}, (across column) covariance matrix.
}
}

