% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmnr-predict.R
\name{bmnr-predict}
\alias{bmnr-predict}
\alias{predict.bmnrfit}
\alias{extract_posterior_predictive_pars}
\alias{extract_posterior_predictive_pars.bmnrfit_mvrnorm}
\alias{extract_posterior_predictive_pars.bmnrfit}
\title{Prediction of \code{bmnr} Models}
\usage{
\method{predict}{bmnrfit}(object, new_data, ..., checkSymmetry = FALSE, type = NULL)

extract_posterior_predictive_pars(object, new_data)

\method{extract_posterior_predictive_pars}{bmnrfit_mvrnorm}(object, new_data)

\method{extract_posterior_predictive_pars}{bmnrfit}(object, new_data)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmnrfit}}.}

\item{new_data}{data frame. Data used to create a design matrices.
Setting \code{new_data = NULL} utilises the pre-trained data (special case).}

\item{...}{These dots are for future extensions and must be empty.}

\item{checkSymmetry}{logical; if \code{FALSE}, skip checking whether the
   covariance matrix is symmetric or not. This will speed up the
   computation but may cause unexpected outputs when ill-behaved
   \code{sigma} is provided. The default value is \code{TRUE}.}

\item{type}{choice. Return type, see \strong{Value} section for more information}
}
\value{
The helper \code{extract_posterior_predictive_pars()} returns a list of the
distribution parameters in the form of a \verb{$flattened_mean} and \verb{$covar}.
Output of the \code{predict} method is dictated by the \code{type} argument:
\itemize{
\item \code{summary}

Tibble representing posterior predictive summaries;
one row per supplied observation from \code{new_data}.
\item \code{numeric}

Tibble with raw MCMC objects stored in a single named list column;
one row per supplied observation from \code{new_data}.
\item \code{raw}

MCMC array of dimension \code{c(n_iter, n_chains, nrow(new_data), n_y)}.
}
}
\description{
Prediction of \code{bmnr} Models
}
\section{Posterior Predictive Distribution}{

TODO. Convert from handwritten notes to Markdown
}

