% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmnr-class.R
\docType{class}
\name{bmnrfit-class}
\alias{bmnrfit-class}
\alias{bmnrfit}
\alias{bmnrfit_simstudy}
\alias{bmnrfit_class}
\alias{bmnrfit_mvrnorm-class}
\alias{bmnrfit_mvrnorm_class}
\alias{bmnrfit_simstudy-class}
\alias{bmnrfit_simstudy_class}
\alias{bmnrfit_mvrnorm_simstudy-class}
\alias{bmnrfit_mvrnorm_simstudy_class}
\title{S4 Class Objects for \code{bmnr}}
\description{
S4 object, inheriting from \code{\link[rstan:stanfit-class]{stanfit}}, that is
returned by \code{\link{bmnr}}.
}
\section{Functions}{
\itemize{
\item \code{bmnrfit_class}: General purpose S4 class (parent class is \code{stanfit}).

\item \code{bmnrfit_mvrnorm_class}: Special case of \code{bmnrfit}, rows assumed independent.

\item \code{bmnrfit_simstudy_class}: Simulation study subclass (parent class is \code{bmnrfit}).

\item \code{bmnrfit_mvrnorm_simstudy_class}: Simulation study subclass
(parent class is \code{bmnrfit_mvrnorm}).

}}
\section{Slots}{

\describe{
\item{\code{data}}{data frame(s). Originally passed data.}

\item{\code{formula}}{list.
* \verb{$model} describes \eqn{Y} and \eqn{X} variables.
* \verb{$coord} describes \eqn{s} coordinate variables.}

\item{\code{prior}}{list. Originally passed prior hyperparameters.}

\item{\code{params}}{list. Underlying parameter values, used in plotting.}
}}

