% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmnr-extract.R
\name{extract-pars}
\alias{extract-pars}
\alias{extract_regr}
\alias{extract_covar_y}
\alias{extract_gp_length}
\alias{extract_linpred}
\alias{extract_covar_s}
\alias{extract_pointwise_log_lik}
\title{Extract Model Parameters}
\usage{
extract_regr(object)

extract_covar_y(object)

extract_gp_length(object)

extract_linpred(object, new_data)

extract_covar_s(object, new_data, .nugget = 1e-08)

extract_pointwise_log_lik(object, new_data)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmnrfit}}.}

\item{new_data}{data frame. Data used to create a design matrices.
Setting \code{new_data = NULL} utilises the pre-trained data (special case).}

\item{.nugget}{numeric. Added to main diagonal for computation stability.}
}
\description{
Extracts or calculates key model parameters and derived quantities
}
