% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp-matern32-cov-ard.R
\name{gp_matern32_cov_ard}
\alias{gp_matern32_cov_ard}
\title{Matern 3/2 Kernel with Characteristic Length Scales}
\usage{
gp_matern32_cov_ard(x_r, gp_scale, gp_length, x_c = NULL)
}
\arguments{
\item{x_r, x_c}{matrix. Each row is a vector input to the kernel.}

\item{gp_scale}{real. Multiplicative amplitude of the kernel function.}

\item{gp_length}{vector. Multiple length scales, one for each dimension.}
}
\value{
Covariance matrix with 'nrow(x)' rows and columns
}
\description{
Matern 3/2 Kernel with Characteristic Length Scales
}
\note{
Exposed stan function.
}
